feat: move plotting methods into `Table`
