--- src/safeds/plotting/__init__.py
@@ -1,3 +1,0 @@
-from ._correlation_heatmap import correlation_heatmap
-from ._lineplot import lineplot
-from ._scatterplot import scatterplot
