+++ tests/safeds/data/tabular/containers/_table/test_lineplot.py
@@ -2,18 +2,17 @@ import _pytest
 import matplotlib.pyplot as plt
 import pandas as pd
 import pytest
-from safeds import plotting
 from safeds.data.tabular.containers import Table
 from safeds.exceptions import UnknownColumnNameError
 
 
-def test_scatterplot(monkeypatch: _pytest.monkeypatch) -> None:
+def test_lineplot(monkeypatch: _pytest.monkeypatch) -> None:
     monkeypatch.setattr(plt, "show", lambda: None)
     table = Table(pd.DataFrame(data={"A": [1, 2, 3], "B": [2, 4, 7]}))
-    plotting.scatterplot(table, "A", "B")
+    table.lineplot("A", "B")
 
 
-def test_scatterplot_wrong_column_name() -> None:
+def test_lineplot_wrong_column_name() -> None:
     with pytest.raises(UnknownColumnNameError):
         table = Table(pd.DataFrame(data={"A": [1, 2, 3], "B": [2, 4, 7]}))
-        plotting.scatterplot(table, "C", "A")
+        table.lineplot("C", "A")
